+++ VisInfProc/cuda_helper.h
@@ -8,11 +8,11 @@
 #define gpuErrchk(ans) { gpuAssert((ans), __FILE__, __LINE__); }
 inline void gpuAssert(cudaError_t code, const char *file, int line, bool abort=true)
 {
-   if (code != cudaSuccess)
-   {
-      fprintf(stderr,"GPUassert: %s %s %d\n", cudaGetErrorString(code), file, line);
-      if (abort) exit(code);
-   }
+    if (code != cudaSuccess)
+        {
+            fprintf(stderr,"GPUassert: %s %s %d\n", cudaGetErrorString(code), file, line);
+            if (abort) exit(code);
+        }
 }
 
 #define gpuKernelErrorCheck() { cudaDeviceSynchronize();\
@@ -23,12 +23,11 @@ inline void gpuAssert(cudaError_t code, const char *file, int line, bool abort=t
        exit(-1);\
      }\
 }
-extern void* cudaCreateBuffer(long size);
-extern void cudaUploadBuffer(void* cpuBuffPtr, void * gpuBuffPtr,long size,cudaStream_t stream = 0);
-extern void cudaDownloadBuffer(void* gpuBuffPtr, void * cpuBuffPtr,long size,cudaStream_t stream = 0);
+extern void* cudaCreateBuffer(size_t size);
+extern void cudaUploadBuffer(void* cpuBuffPtr, void * gpuBuffPtr,size_t size,cudaStream_t stream = 0);
+extern void cudaDownloadBuffer(void* gpuBuffPtr, void * cpuBuffPtr,size_t size,cudaStream_t stream = 0);
 extern void cudaFreeBuffer(void* gpuBuffPtr);
-extern void cudaCopyBuffer(void* gpuBuffPtrDest, void * gpuBuffPtrSrc,long size,cudaStream_t stream = 0);
-extern void cudaSetDoubleBuffer(float* gpuBuffPtr,float v, long size,cudaStream_t stream = 0);
+extern void cudaCopyBuffer(void* gpuBuffPtrDest, void * gpuBuffPtrSrc,size_t size,cudaStream_t stream = 0);
+extern void cudaSetDoubleBuffer(float* gpuBuffPtr, float v, size_t size, cudaStream_t stream = 0);
 
 #endif // CUDA_HELPER_H
-
